requirements.txt
RPi.GPIO
adafruit-circuitpython-bme280
pandas
